+++ inject/inject-thrift-client/src/test/scala/com/twitter/inject/thrift/DoEverythingThriftClientModuleFeatureTest.scala
@@ -114,8 +114,6 @@ class DoEverythingThriftClientModuleFeatureTest extends FeatureTest with HttpTes
     )
 
     httpServer1.httpGet(path = "/echo?msg=Bob", andExpect = Ok, withBody = "BobBobBob")
-    httpServer1.inMemoryStats.stats.assert("route/config/POST/response_size", Seq(1.0f, 1.0f))
-    httpServer1.inMemoryStats.stats.assert("route/echo/GET/response_size", Seq(9.0f))
   }
 
   // test EchoService.FutureIface
@@ -138,8 +136,6 @@ class DoEverythingThriftClientModuleFeatureTest extends FeatureTest with HttpTes
     )
 
     httpServer2.httpGet(path = "/echo?msg=Bob", andExpect = Ok, withBody = "BobBobBob")
-    httpServer2.inMemoryStats.stats.assert("route/config/POST/response_size", Seq(1.0f, 1.0f))
-    httpServer2.inMemoryStats.stats.assert("route/echo/GET/response_size", Seq(9.0f))
   }
 
   // test EchoService.MethodPerEndpoint
@@ -162,8 +158,6 @@ class DoEverythingThriftClientModuleFeatureTest extends FeatureTest with HttpTes
     )
 
     httpServer3.httpGet(path = "/echo?msg=Bob", andExpect = Ok, withBody = "BobBobBob")
-    httpServer3.inMemoryStats.stats.assert("route/config/POST/response_size", Seq(1.0f, 1.0f))
-    httpServer3.inMemoryStats.stats.assert("route/echo/GET/response_size", Seq(9.0f))
   }
 
   test("ThriftClient#asClosable") {